fix(postgresql): cast QueueStatusAdmin enum to int for Npgsql 10.x (#155)
GetQueueCountQueryPrepareHandler bound the QueueStatusAdmin enum directly to
an Int32 parameter. Npgsql 10.x's stricter parameter writer rejects this with
InvalidCastException; pre-10.x and Microsoft.Data.SqlClient tolerated it. Cast
to (int) to match the declared DbType.Int32. Shared relational handler, so this
fixes PostgreSQL, SQL Server, and SQLite.

+++ Source/DotNetWorkQueue.Transport.RelationalDatabase/Basic/QueryPrepareHandler/GetQueueCountQueryPrepareHandler.cs
@@ -50,7 +50,7 @@ namespace DotNetWorkQueue.Transport.RelationalDatabase.Basic.QueryPrepareHandler
                 var status = dbCommand.CreateParameter();
                 status.ParameterName = "@Status";
                 status.DbType = DbType.Int32;
-                status.Value = query.Status.Value;
+                status.Value = (int)query.Status.Value;
                 dbCommand.Parameters.Add(status);
             }
             else
